Update every migration to avoid loss of data

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -27,6 +27,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('users');
+        if (App::isLocal()) {
+            Schema::dropIfExists('users');
+        }
     }
 };
