+++ database/migrations/2019_12_14_000001_create_personal_access_tokens_table.php
@@ -28,6 +28,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('personal_access_tokens');
+        if (App::isLocal()) {
+            Schema::dropIfExists('personal_access_tokens');
+        }
     }
 };
