+++ database/migrations/2023_12_22_202805_create_admin_ips_table.php
@@ -23,6 +23,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('admin_ips');
+        if (App::isLocal()) {
+            Schema::dropIfExists('admin_ips');
+        }
     }
 };
