+++ database/migrations/admin/2023_04_29_201120_create_blogs_table.php
@@ -27,6 +27,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('blogs');
+        if (App::isLocal()) {
+            Schema::dropIfExists('blogs');
+        }
     }
 };
