+++ database/migrations/admin/2023_05_02_104845_create_profiles_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('profiles');
+        if (App::isLocal()) {
+            Schema::dropIfExists('profiles');
+        }
     }
 };
