+++ database/migrations/admin/2023_05_02_111836_create_skills_table.php
@@ -26,6 +26,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('skills');
+        if (App::isLocal()) {
+            Schema::dropIfExists('skills');
+        }
     }
 };
