+++ database/migrations/orchid/2015_04_12_000000_create_orchid_users_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn(['permissions']);
-        });
+        if (App::isLocal()) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn(['permissions']);
+            });
+        }
     }
 };
