+++ database/migrations/orchid/2015_10_19_214424_create_orchid_roles_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('roles');
+        if (App::isLocal()) {
+            Schema::dropIfExists('roles');
+        }
     }
 };
