+++ database/migrations/orchid/2015_10_19_214425_create_orchid_role_users_table.php
@@ -33,6 +33,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('role_users');
+        if (App::isLocal()) {
+            Schema::dropIfExists('role_users');
+        }
     }
 };
