+++ database/migrations/orchid/2017_09_17_125801_create_notifications_table.php
@@ -28,6 +28,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('notifications');
+        if (App::isLocal()) {
+            Schema::dropIfExists('notifications');
+        }
     }
 };
